backend/requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9

# Security & Encryption
cryptography==41.0.7

# Session Management
Flask-Session==0.5.0
redis==5.0.1

# Acumatica Integration
easy-acumatica>=0.5.10

# Environment & Configuration
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0

# Development Server (optional)
waitress==2.1.2

# Utilities
python-dateutil==2.8.2